@@ -4,13 +4,19 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as cp from 'child_process';
+import * as gulp from 'gulp';
 import * as path from 'path';
+import { gulp_installAzureAccount, gulp_webpack } from 'vscode-azureextensiondev';
 
 const env = process.env;
 
-function test() {
+function test(): cp.ChildProcess {
     env.DEBUGTELEMETRY = '1';
+    env.CODE_TESTS_WORKSPACE = path.join(__dirname, 'test/test.code-workspace');
+    env.CODE_TESTS_PATH = path.join(__dirname, 'dist/test');
     return cp.spawn('node', ['./node_modules/vscode/bin/test'], { stdio: 'inherit', env });
 }
 
-exports.test = test;
+exports['webpack-dev'] = () => gulp_webpack('development');
+exports['webpack-prod'] = () => gulp_webpack('production');
+exports.test = gulp.series(gulp_installAzureAccount, test);
