@@ -5,11 +5,7 @@
 // tslint:disable:no-unused-expression
 
 import * as assert from "assert";
-
-import * as Json from "../src/JSON";
-import * as language from "../src/Language";
-
-import { ParameterDefinition } from "../src/ParameterDefinition";
+import { Json, Language, ParameterDefinition } from "../extension.bundle";
 
 suite("ParameterDefinition", () => {
     suite("constructor(Json.Property)", () => {
@@ -22,13 +18,13 @@ suite("ParameterDefinition", () => {
         });
 
         test("with property with no metadata", () => {
-            const parameterName = new Json.StringValue(new language.Span(0, 13), "parameterName");
-            const parameterDefinition = new Json.ObjectValue(new language.Span(16, 2), []);
+            const parameterName = new Json.StringValue(new Language.Span(0, 13), "parameterName");
+            const parameterDefinition = new Json.ObjectValue(new Language.Span(16, 2), []);
             const property = new Json.Property(parameterName.span.union(parameterDefinition.span), parameterName, parameterDefinition);
             const pd = new ParameterDefinition(property);
             assert.deepStrictEqual(pd.name, parameterName);
             assert.deepStrictEqual(pd.description, null);
-            assert.deepStrictEqual(pd.span, new language.Span(0, 18));
+            assert.deepStrictEqual(pd.span, new Language.Span(0, 18));
         });
     });
 });
