@@ -7,17 +7,7 @@
 // tslint:disable:promise-function-async
 
 import * as assert from "assert";
-
-import * as Json from "../src/JSON";
-import * as language from "../src/Language";
-import * as Reference from "../src/Reference";
-import * as TLE from "../src/TLE";
-
-import { AzureRMAssets, FunctionMetadata, FunctionsMetadata } from "../src/AzureRMAssets";
-import { DeploymentTemplate } from "../src/DeploymentTemplate";
-import { IncorrectArgumentsCountIssue } from "../src/IncorrectArgumentsCountIssue";
-import { PositionContext } from "../src/PositionContext";
-import { UnrecognizedFunctionIssue } from "../src/UnrecognizedFunctionIssue";
+import { AzureRMAssets, DeploymentTemplate, FunctionMetadata, FunctionsMetadata, IncorrectArgumentsCountIssue, Language, PositionContext, Reference, TLE, UnrecognizedFunctionIssue } from "../extension.bundle";
 
 suite("TLE", () => {
     suite("StringValue", () => {
@@ -170,14 +160,14 @@ suite("TLE", () => {
                 let source = new TLE.NumberValue(TLE.Token.createNumber(5, "2"));
                 let leftSquareBracket = TLE.Token.createLeftSquareBracket(6);
                 let arrayAccess = new TLE.ArrayAccessValue(source, leftSquareBracket, null, null);
-                assert.deepStrictEqual(new language.Span(5, 2), arrayAccess.getSpan());
+                assert.deepStrictEqual(new Language.Span(5, 2), arrayAccess.getSpan());
             });
 
             test("with whitespace between source and left square bracket", () => {
                 let source = new TLE.NumberValue(TLE.Token.createNumber(5, "2"));
                 let leftSquareBracket = TLE.Token.createLeftSquareBracket(8);
                 let arrayAccess = new TLE.ArrayAccessValue(source, leftSquareBracket, null, null);
-                assert.deepStrictEqual(new language.Span(5, 4), arrayAccess.getSpan());
+                assert.deepStrictEqual(new Language.Span(5, 4), arrayAccess.getSpan());
             });
 
             test("with no right square bracket", () => {
@@ -185,7 +175,7 @@ suite("TLE", () => {
                 let leftSquareBracket = TLE.Token.createLeftSquareBracket(8);
                 let index = new TLE.NumberValue(TLE.Token.createNumber(10, "10"));
                 let arrayAccess = new TLE.ArrayAccessValue(source, leftSquareBracket, index, null);
-                assert.deepStrictEqual(new language.Span(5, 7), arrayAccess.getSpan());
+                assert.deepStrictEqual(new Language.Span(5, 7), arrayAccess.getSpan());
             });
 
             test("with no index", () => {
@@ -193,7 +183,7 @@ suite("TLE", () => {
                 let leftSquareBracket = TLE.Token.createLeftSquareBracket(8);
                 let rightSquareBracket = TLE.Token.createRightSquareBracket(12);
                 let arrayAccess = new TLE.ArrayAccessValue(source, leftSquareBracket, null, rightSquareBracket);
-                assert.deepStrictEqual(new language.Span(5, 8), arrayAccess.getSpan());
+                assert.deepStrictEqual(new Language.Span(5, 8), arrayAccess.getSpan());
             });
 
             test("with complete array access", () => {
@@ -202,7 +192,7 @@ suite("TLE", () => {
                 let index = new TLE.NumberValue(TLE.Token.createNumber(10, "10"));
                 let rightSquareBracket = TLE.Token.createRightSquareBracket(12);
                 let arrayAccess = new TLE.ArrayAccessValue(source, leftSquareBracket, index, rightSquareBracket);
-                assert.deepStrictEqual(new language.Span(5, 8), arrayAccess.getSpan());
+                assert.deepStrictEqual(new Language.Span(5, 8), arrayAccess.getSpan());
             });
         });
 
@@ -293,43 +283,43 @@ suite("TLE", () => {
             test("with name", () => {
                 let f = TLE.Parser.parse("\"[concat]\"").expression;
                 assert(f instanceof TLE.FunctionValue);
-                assert.deepStrictEqual(new language.Span(2, 6), f.getSpan());
+                assert.deepStrictEqual(new Language.Span(2, 6), f.getSpan());
             });
 
             test("with left parenthesis", () => {
                 let f = TLE.Parser.parse("\"[concat(]\"").expression;
                 assert(f instanceof TLE.FunctionValue);
-                assert.deepStrictEqual(new language.Span(2, 7), f.getSpan());
+                assert.deepStrictEqual(new Language.Span(2, 7), f.getSpan());
             });
 
             test("with one argument and no right parenthesis", () => {
                 let f = TLE.Parser.parse("\"[concat(70").expression;
                 assert(f instanceof TLE.FunctionValue);
-                assert.deepStrictEqual(new language.Span(2, 9), f.getSpan());
+                assert.deepStrictEqual(new Language.Span(2, 9), f.getSpan());
             });
 
             test("with two arguments and no right parenthesis", () => {
                 let f = TLE.Parser.parse("\"[concat(70, 3").expression;
                 assert(f instanceof TLE.FunctionValue);
-                assert.deepStrictEqual(new language.Span(2, 12), f.getSpan());
+                assert.deepStrictEqual(new Language.Span(2, 12), f.getSpan());
             });
 
             test("with left and right parenthesis and no arguments", () => {
                 let f = TLE.Parser.parse("\"[concat()\"").expression;
                 assert(f instanceof TLE.FunctionValue);
-                assert.deepStrictEqual(new language.Span(2, 8), f.getSpan());
+                assert.deepStrictEqual(new Language.Span(2, 8), f.getSpan());
             });
 
             test("with left and right parenthesis and arguments", () => {
                 let f = TLE.Parser.parse("\"[concat('hello', 'world')\"").expression;
                 assert(f instanceof TLE.FunctionValue);
-                assert.deepStrictEqual(new language.Span(2, 24), f.getSpan());
+                assert.deepStrictEqual(new Language.Span(2, 24), f.getSpan());
             });
 
             test("with last argument missing and no right parenthesis", () => {
                 let f = TLE.Parser.parse("\"[concat('hello',").expression;
                 assert(f instanceof TLE.FunctionValue);
-                assert.deepStrictEqual(new language.Span(2, 15), f.getSpan());
+                assert.deepStrictEqual(new Language.Span(2, 15), f.getSpan());
             });
         });
     });
@@ -427,7 +417,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     visitor.errors,
                     [
-                        new language.Issue(new language.Span(17, 2), "Undefined parameter reference: ''")
+                        new Language.Issue(new Language.Span(17, 2), "Undefined parameter reference: ''")
                     ]
                 );
             });
@@ -448,7 +438,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     visitor.errors,
                     [
-                        new language.Issue(new language.Span(17, 2), "Undefined variable reference: ''")
+                        new Language.Issue(new Language.Span(17, 2), "Undefined variable reference: ''")
                     ]
                 );
             });
@@ -510,8 +500,8 @@ suite("TLE", () => {
                 assert.equal(null, pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(2, 1), "Expected a right square bracket (']')."),
-                        new language.Issue(new language.Span(1, 1), "Expected a function or property expression.")
+                        new Language.Issue(new Language.Span(2, 1), "Expected a right square bracket (']')."),
+                        new Language.Issue(new Language.Span(1, 1), "Expected a function or property expression.")
                     ],
                     pr.errors);
             });
@@ -556,7 +546,7 @@ suite("TLE", () => {
                 assert.equal(null, pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(2), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
-                    [new language.Issue(new language.Span(1, 2), "Expected a function or property expression.")],
+                    [new Language.Issue(new Language.Span(1, 2), "Expected a function or property expression.")],
                     pr.errors);
             });
 
@@ -567,7 +557,7 @@ suite("TLE", () => {
                 assert.equal(null, pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(4), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
-                    [new language.Issue(new language.Span(3, 2), "Expected a function or property expression.")],
+                    [new Language.Issue(new Language.Span(3, 2), "Expected a function or property expression.")],
                     pr.errors);
             });
 
@@ -578,7 +568,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(null, pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(6), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
-                    [new language.Issue(new language.Span(1, 6), "Expected a function or property expression.")],
+                    [new Language.Issue(new Language.Span(1, 6), "Expected a function or property expression.")],
                     pr.errors);
             });
 
@@ -601,8 +591,8 @@ suite("TLE", () => {
                 assert.equal(null, pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(2, 6), "Missing function argument list."),
-                        new language.Issue(new language.Span(8, 1), "Expected a right square bracket (']').")
+                        new Language.Issue(new Language.Span(2, 6), "Missing function argument list."),
+                        new Language.Issue(new Language.Span(8, 1), "Expected a right square bracket (']').")
                     ],
                     pr.errors);
             });
@@ -614,7 +604,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(new TLE.FunctionValue(TLE.Token.createLiteral(2, "concat"), null, [], [], null), pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(8), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
-                    [new language.Issue(new language.Span(2, 6), "Missing function argument list.")],
+                    [new Language.Issue(new Language.Span(2, 6), "Missing function argument list.")],
                     pr.errors);
             });
 
@@ -633,8 +623,8 @@ suite("TLE", () => {
                 assert.equal(null, pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(9, 1), "Expected a right parenthesis (')')."),
-                        new language.Issue(new language.Span(10, 1), "Expected a right square bracket (']').")
+                        new Language.Issue(new Language.Span(9, 1), "Expected a right parenthesis (')')."),
+                        new Language.Issue(new Language.Span(10, 1), "Expected a right square bracket (']').")
                     ],
                     pr.errors);
             });
@@ -653,7 +643,7 @@ suite("TLE", () => {
                     pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(10), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
-                    [new language.Issue(new language.Span(10, 1), "Expected a right parenthesis (')').")],
+                    [new Language.Issue(new Language.Span(10, 1), "Expected a right parenthesis (')').")],
                     pr.errors);
             });
 
@@ -672,8 +662,8 @@ suite("TLE", () => {
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(9), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(8, 1), "Expected the end of the string."),
-                        new language.Issue(new language.Span(2, 6), "Missing function argument list."),
+                        new Language.Issue(new Language.Span(8, 1), "Expected the end of the string."),
+                        new Language.Issue(new Language.Span(2, 6), "Missing function argument list."),
                     ],
                     pr.errors);
             });
@@ -730,7 +720,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(null, pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(9, 1), "Expected a right square bracket (']').")
+                        new Language.Issue(new Language.Span(9, 1), "Expected a right square bracket (']').")
                     ],
                     pr.errors);
             });
@@ -750,7 +740,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(null, pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(10, 1), "Expected a right square bracket (']').")
+                        new Language.Issue(new Language.Span(10, 1), "Expected a right square bracket (']').")
                     ],
                     pr.errors);
             });
@@ -804,7 +794,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(16, 1), "Expected a constant string, function, or property expression.")
+                        new Language.Issue(new Language.Span(16, 1), "Expected a constant string, function, or property expression.")
                     ]);
 
                 const concat: TLE.FunctionValue = TLE.asFunctionValue(pr.expression);
@@ -835,7 +825,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(9, 1), "Expected a constant string, function, or property expression.")
+                        new Language.Issue(new Language.Span(9, 1), "Expected a constant string, function, or property expression.")
                     ]);
 
                 const concat: TLE.FunctionValue = TLE.asFunctionValue(pr.expression);
@@ -866,8 +856,8 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(13, 1), "Expected a constant string, function, or property expression."),
-                        new language.Issue(new language.Span(13, 1), "Expected a right square bracket (']').")
+                        new Language.Issue(new Language.Span(13, 1), "Expected a constant string, function, or property expression."),
+                        new Language.Issue(new Language.Span(13, 1), "Expected a right square bracket (']').")
                     ]);
 
                 const concat: TLE.FunctionValue = TLE.asFunctionValue(pr.expression);
@@ -909,9 +899,9 @@ suite("TLE", () => {
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(12), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(9, 1), "Expected a constant string, function, or property expression."),
-                        new language.Issue(new language.Span(10, 1), "Expected a constant string, function, or property expression."),
-                        new language.Issue(new language.Span(11, 1), "Expected a constant string, function, or property expression.")
+                        new Language.Issue(new Language.Span(9, 1), "Expected a constant string, function, or property expression."),
+                        new Language.Issue(new Language.Span(10, 1), "Expected a constant string, function, or property expression."),
+                        new Language.Issue(new Language.Span(11, 1), "Expected a constant string, function, or property expression.")
                     ],
                     pr.errors);
             });
@@ -1037,8 +1027,8 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(9, 3), "A constant string is missing an end quote."),
-                        new language.Issue(new language.Span(12, 1), "Expected a right square bracket (']').")
+                        new Language.Issue(new Language.Span(9, 3), "A constant string is missing an end quote."),
+                        new Language.Issue(new Language.Span(12, 1), "Expected a right square bracket (']').")
                     ]);
 
                 const concat: TLE.FunctionValue = TLE.asFunctionValue(pr.expression);
@@ -1063,7 +1053,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(16, 2), "Expected a comma (',')."),
+                        new Language.Issue(new Language.Span(16, 2), "Expected a comma (',')."),
                     ]);
 
                 const concat: TLE.FunctionValue = TLE.asFunctionValue(pr.expression);
@@ -1088,8 +1078,8 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(16, 2), "Expected a comma (',')."),
-                        new language.Issue(new language.Span(18, 7), "Expected a comma (',')."),
+                        new Language.Issue(new Language.Span(16, 2), "Expected a comma (',')."),
+                        new Language.Issue(new Language.Span(18, 7), "Expected a comma (',')."),
                     ]);
 
                 const concat: TLE.FunctionValue = TLE.asFunctionValue(pr.expression);
@@ -1141,7 +1131,7 @@ suite("TLE", () => {
                         TLE.Token.createRightParenthesis(16)),
                     pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(21), pr.rightSquareBracketToken);
-                assert.deepStrictEqual([new language.Issue(new language.Span(17, 4), "Expected the end of the string.")], pr.errors);
+                assert.deepStrictEqual([new Language.Issue(new Language.Span(17, 4), "Expected the end of the string.")], pr.errors);
             });
 
             test("with quoted string instead of literal for property access", () => {
@@ -1152,7 +1142,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(18, 6), "Expected a literal value.")
+                        new Language.Issue(new Language.Span(18, 6), "Expected a literal value.")
                     ]);
 
                 const propertyAccess: TLE.PropertyAccess = TLE.asPropertyAccessValue(pr.expression);
@@ -1178,7 +1168,7 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(18, 1), "Expected a literal value.")
+                        new Language.Issue(new Language.Span(18, 1), "Expected a literal value.")
                     ]);
 
                 const propertyAccess: TLE.PropertyAccess = TLE.asPropertyAccessValue(pr.expression);
@@ -1344,7 +1334,7 @@ suite("TLE", () => {
                         TLE.Token.createRightParenthesis(8)),
                     pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(16), pr.rightSquareBracketToken);
-                assert.deepStrictEqual([new language.Issue(new language.Span(9, 7), "Expected the end of the string.")], pr.errors);
+                assert.deepStrictEqual([new Language.Issue(new Language.Span(9, 7), "Expected the end of the string.")], pr.errors);
             });
 
             test("with function after string", () => {
@@ -1360,7 +1350,7 @@ suite("TLE", () => {
                         TLE.Token.createRightParenthesis(15)),
                     pr.expression);
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(16), pr.rightSquareBracketToken);
-                assert.deepStrictEqual([new language.Issue(new language.Span(7, 7), "Expected the end of the string.")], pr.errors);
+                assert.deepStrictEqual([new Language.Issue(new Language.Span(7, 7), "Expected the end of the string.")], pr.errors);
             });
 
             test("with string followed by literal", () => {
@@ -1378,8 +1368,8 @@ suite("TLE", () => {
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(14), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(2, 7), "Expected a literal value."),
-                        new language.Issue(new language.Span(9, 5), "Missing function argument list."),
+                        new Language.Issue(new Language.Span(2, 7), "Expected a literal value."),
+                        new Language.Issue(new Language.Span(9, 5), "Missing function argument list."),
                     ],
                     pr.errors);
             });
@@ -1399,8 +1389,8 @@ suite("TLE", () => {
                 assert.deepStrictEqual(TLE.Token.createRightSquareBracket(14), pr.rightSquareBracketToken);
                 assert.deepStrictEqual(
                     [
-                        new language.Issue(new language.Span(7, 7), "Expected the end of the string."),
-                        new language.Issue(new language.Span(2, 5), "Missing function argument list."),
+                        new Language.Issue(new Language.Span(7, 7), "Expected the end of the string."),
+                        new Language.Issue(new Language.Span(2, 5), "Missing function argument list."),
                     ],
                     pr.errors);
             });
@@ -1410,9 +1400,9 @@ suite("TLE", () => {
                 assert.deepStrictEqual(
                     pr.errors,
                     [
-                        new language.Issue(new language.Span(116, 1), "Nothing should exist after the closing ']' except for whitespace."),
-                        new language.Issue(new language.Span(118, 1), "Nothing should exist after the closing ']' except for whitespace."),
-                        new language.Issue(new language.Span(119, 1), "Nothing should exist after the closing ']' except for whitespace.")
+                        new Language.Issue(new Language.Span(116, 1), "Nothing should exist after the closing ']' except for whitespace."),
+                        new Language.Issue(new Language.Span(118, 1), "Nothing should exist after the closing ']' except for whitespace."),
+                        new Language.Issue(new Language.Span(119, 1), "Nothing should exist after the closing ']' except for whitespace.")
                     ]);
             });
         });
@@ -1423,21 +1413,21 @@ suite("TLE", () => {
             test("Negative startIndex", () => {
                 let t = TLE.Token.createLeftParenthesis(-1);
                 assert.equal(TLE.TokenType.LeftParenthesis, t.getType());
-                assert.deepStrictEqual(new language.Span(-1, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(-1, 1), t.span);
                 assert.equal("(", t.stringValue);
             });
 
             test("Zero startIndex", () => {
                 let t = TLE.Token.createLeftParenthesis(0);
                 assert.equal(TLE.TokenType.LeftParenthesis, t.getType());
-                assert.deepStrictEqual(new language.Span(0, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(0, 1), t.span);
                 assert.equal("(", t.stringValue);
             });
 
             test("Positive startIndex", () => {
                 let t = TLE.Token.createLeftParenthesis(7);
                 assert.equal(TLE.TokenType.LeftParenthesis, t.getType());
-                assert.deepStrictEqual(new language.Span(7, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(7, 1), t.span);
                 assert.equal("(", t.stringValue);
             });
         });
@@ -1446,21 +1436,21 @@ suite("TLE", () => {
             test("Negative startIndex", () => {
                 let t = TLE.Token.createRightParenthesis(-1);
                 assert.equal(TLE.TokenType.RightParenthesis, t.getType());
-                assert.deepStrictEqual(new language.Span(-1, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(-1, 1), t.span);
                 assert.equal(")", t.stringValue);
             });
 
             test("Zero startIndex", () => {
                 let t = TLE.Token.createRightParenthesis(0);
                 assert.equal(TLE.TokenType.RightParenthesis, t.getType());
-                assert.deepStrictEqual(new language.Span(0, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(0, 1), t.span);
                 assert.equal(")", t.stringValue);
             });
 
             test("Positive startIndex", () => {
                 let t = TLE.Token.createRightParenthesis(7);
                 assert.equal(TLE.TokenType.RightParenthesis, t.getType());
-                assert.deepStrictEqual(new language.Span(7, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(7, 1), t.span);
                 assert.equal(")", t.stringValue);
             });
         });
@@ -1469,21 +1459,21 @@ suite("TLE", () => {
             test("Negative startIndex", () => {
                 let t = TLE.Token.createLeftSquareBracket(-1);
                 assert.equal(TLE.TokenType.LeftSquareBracket, t.getType());
-                assert.deepStrictEqual(new language.Span(-1, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(-1, 1), t.span);
                 assert.equal("[", t.stringValue);
             });
 
             test("Zero startIndex", () => {
                 let t = TLE.Token.createLeftSquareBracket(0);
                 assert.equal(TLE.TokenType.LeftSquareBracket, t.getType());
-                assert.deepStrictEqual(new language.Span(0, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(0, 1), t.span);
                 assert.equal("[", t.stringValue);
             });
 
             test("Positive startIndex", () => {
                 let t = TLE.Token.createLeftSquareBracket(7);
                 assert.equal(TLE.TokenType.LeftSquareBracket, t.getType());
-                assert.deepStrictEqual(new language.Span(7, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(7, 1), t.span);
                 assert.equal("[", t.stringValue);
             });
         });
@@ -1492,21 +1482,21 @@ suite("TLE", () => {
             test("Negative startIndex", () => {
                 let t = TLE.Token.createRightSquareBracket(-1);
                 assert.equal(TLE.TokenType.RightSquareBracket, t.getType());
-                assert.deepStrictEqual(new language.Span(-1, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(-1, 1), t.span);
                 assert.equal("]", t.stringValue);
             });
 
             test("Zero startIndex", () => {
                 let t = TLE.Token.createRightSquareBracket(0);
                 assert.equal(TLE.TokenType.RightSquareBracket, t.getType());
-                assert.deepStrictEqual(new language.Span(0, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(0, 1), t.span);
                 assert.equal("]", t.stringValue);
             });
 
             test("Positive startIndex", () => {
                 let t = TLE.Token.createRightSquareBracket(7);
                 assert.equal(TLE.TokenType.RightSquareBracket, t.getType());
-                assert.deepStrictEqual(new language.Span(7, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(7, 1), t.span);
                 assert.equal("]", t.stringValue);
             });
         });
@@ -1515,21 +1505,21 @@ suite("TLE", () => {
             test("Negative startIndex", () => {
                 let t = TLE.Token.createComma(-1);
                 assert.equal(TLE.TokenType.Comma, t.getType());
-                assert.deepStrictEqual(new language.Span(-1, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(-1, 1), t.span);
                 assert.equal(",", t.stringValue);
             });
 
             test("Zero startIndex", () => {
                 let t = TLE.Token.createComma(0);
                 assert.equal(TLE.TokenType.Comma, t.getType());
-                assert.deepStrictEqual(new language.Span(0, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(0, 1), t.span);
                 assert.equal(",", t.stringValue);
             });
 
             test("Positive startIndex", () => {
                 let t = TLE.Token.createComma(7);
                 assert.equal(TLE.TokenType.Comma, t.getType());
-                assert.deepStrictEqual(new language.Span(7, 1), t.span);
+                assert.deepStrictEqual(new Language.Span(7, 1), t.span);
                 assert.equal(",", t.stringValue);
             });
         });
@@ -1780,7 +1770,7 @@ suite("TLE", () => {
                 assert(visitor);
                 assert.deepStrictEqual(
                     [
-                        new UnrecognizedFunctionIssue(new language.Span(2, 11), "concatenate")
+                        new UnrecognizedFunctionIssue(new Language.Span(2, 11), "concatenate")
                     ],
                     visitor.errors);
                 assert.equal(visitor.errors[0].message, "Unrecognized function name 'concatenate'.");
@@ -1855,7 +1845,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 5), "The function 'add' takes 2 arguments.", "add", 0, 2, 2)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 5), "The function 'add' takes 2 arguments.", "add", 0, 2, 2)]);
                     });
             });
 
@@ -1867,7 +1857,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 6), "The function 'add' takes 2 arguments.", "add", 1, 2, 2)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 6), "The function 'add' takes 2 arguments.", "add", 1, 2, 2)]);
                     });
             });
 
@@ -1889,7 +1879,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 12), "The function 'add' takes 2 arguments.", "add", 3, 2, 2)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 12), "The function 'add' takes 2 arguments.", "add", 3, 2, 2)]);
                     });
             });
 
@@ -1901,7 +1891,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 12), "The function 'add' takes 2 arguments.", "add", 3, 2, 2)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 12), "The function 'add' takes 2 arguments.", "add", 3, 2, 2)]);
                     });
             });
 
@@ -1913,7 +1903,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 12), "The function 'resourceId' takes at least 2 arguments.", "resourceId", 0, 2, null)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 12), "The function 'resourceId' takes at least 2 arguments.", "resourceId", 0, 2, null)]);
                     });
             });
 
@@ -1925,7 +1915,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 13), "The function 'resourceId' takes at least 2 arguments.", "resourceId", 1, 2, null)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 13), "The function 'resourceId' takes at least 2 arguments.", "resourceId", 1, 2, null)]);
                     });
             });
 
@@ -1957,7 +1947,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 11), "The function 'substring' takes between 1 and 3 arguments.", "substring", 0, 1, 3)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 11), "The function 'substring' takes between 1 and 3 arguments.", "substring", 0, 1, 3)]);
                     });
             });
 
@@ -1999,7 +1989,7 @@ suite("TLE", () => {
                         assert(visitor);
                         assert.deepStrictEqual(
                             visitor.errors,
-                            [new IncorrectArgumentsCountIssue(new language.Span(2, 30), "The function 'substring' takes between 1 and 3 arguments.", "substring", 4, 1, 3)]);
+                            [new IncorrectArgumentsCountIssue(new Language.Span(2, 30), "The function 'substring' takes between 1 and 3 arguments.", "substring", 4, 1, 3)]);
                     });
             });
         });
@@ -2027,7 +2017,7 @@ suite("TLE", () => {
                 const visitor = TLE.UndefinedVariablePropertyVisitor.visit(context.tleValue, dt);
                 assert.deepStrictEqual(
                     visitor.errors,
-                    [new language.Issue(new language.Span(18, 6), `Property "apples" is not a defined property of "variables('v1')".`)]);
+                    [new Language.Issue(new Language.Span(18, 6), `Property "apples" is not a defined property of "variables('v1')".`)]);
             });
 
             test("with grandchild property access from variable reference to non-object variable", () => {
@@ -2036,7 +2026,7 @@ suite("TLE", () => {
                 const visitor = TLE.UndefinedVariablePropertyVisitor.visit(context.tleValue, dt);
                 assert.deepStrictEqual(
                     visitor.errors,
-                    [new language.Issue(new language.Span(18, 6), `Property "apples" is not a defined property of "variables('v1')".`)]);
+                    [new Language.Issue(new Language.Span(18, 6), `Property "apples" is not a defined property of "variables('v1')".`)]);
             });
         });
     });
@@ -2061,7 +2051,7 @@ suite("TLE", () => {
                 assert(visitor);
                 assert.deepStrictEqual(
                     visitor.references,
-                    new Reference.List(Reference.ReferenceKind.Parameter, [new language.Span(14, 5)]));
+                    new Reference.List(Reference.ReferenceKind.Parameter, [new Language.Span(14, 5)]));
             });
         });
     });
